Create AutoBoxingAndUnboxing.java
Autoboxing is a process by which a primitive type is automatically encapsulated (boxed) into its equivalent type wrapper whenever an object of that type is needed.
Auto-unboxing is the process by which the value of a boxed object is automatically extracted or unboxed from a type wrapper when its value is needed. No need for calling method intValue( ) and doubleValue( ).