Create SingleMemberAnnotations.java
A single-member annotation contains only one member.
It works like a normal annotation except that it allows a shorthand form of specifying the
value of the member.
When only one member is present, the value for that member can be specified when the annotation is applied. Need not specify the name of the member.